user can only delete their own posts

+++ controllers/wall.js
@@ -22,9 +22,17 @@ module.exports = {
     },
     deletePost: async (req, res) => {
         try {
-            await Post.findOneAndDelete({_id:req.body.postId})
-            console.log('Post deleted')
-            res.json('Post deleted')
+            // Find the post by ID
+            const post = await Post.findById(req.body.postId)
+           // Check if the post exists and if the logged-in user is the owner
+            if (post && post.userId.equals(req.user._id)) {
+                await Post.findOneAndDelete({ _id: req.body.postId })
+                console.log('Post deleted')
+                res.json('Post deleted')
+            } else {
+                console.log('Unauthorized attempt to delete post')
+                res.status(403).json('You are not authorized to delete this post')
+            }
         } catch (error) {
             console.log('failed to delete post on database', error)
         }
